# backend/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MERN API",
    "version": "1.0.0",
    "description": "API documentation for MERN application"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the user"
          },
          "name": {
            "type": "string",
            "description": "The name of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "description": "The password of the user"
          },
          "date": {
            "type": "string",
            "description": "The date the user was created"
          }
        },
        "example": {
          "id": "d5fE_asz",
          "name": "John Doe",
          "email": "johndoe@example.com",
          "password": "123456"
        }
      },
      "Note": {
        "type": "object",
        "required": ["tag", "title", "description"],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the Note"
          },
          "user": {
            "type": "string",
            "description": "The id of the user who created the note"
          },
          "tag": {
            "type": "string",
            "description": "The tag of the Note"
          },
          "title": {
            "type": "string",
            "description": "The title of the Note"
          },
          "description": {
            "type": "string",
            "description": "The description of the Note"
          },
          "date": {
            "type": "string",
            "description": "The date the Note was created"
          }
        },
        "example": {
          "id": "d5fE_asz",
          "user": "d5fE_asz312",
          "tag": "Personal",
          "title": "First Note",
          "description": "This is my very first Note"
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token for authorization"
      }
    }
  },
  "paths": {
    "/api/auth": {
      "get": {
        "summary": "Returns the list of all users",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "The list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/createuser": {
      "post": {
        "summary": "Registers a new user",
        "tags": ["Auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties":{
                  "name": {
                    "type": "string",
                    "description": "The name of the user"
                  },
                  "email": {
                    "type": "string",
                    "description": "The email of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user"
                  }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "response": "User registered successfully"
          },
          "400": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["Auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "The email of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "headers": {
              "Set-Cookie": {
                "description": "set token & refreshToken cookie",
                "schema": {
                  "type": "string",
                  "example": "token=eyJsae1dsa.1e1esafd1dsd.131313dsad1 { httpOnly: true, domain: 'localhost', path: '/' }"
                }
              }
            },
            "content": {
              "application/json": {
                "accessToken": "string",
                "refreshToken": "string"
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/getuser": {
      "post": {
        "summary": "Get Details Of Current User",
        "tags": ["Auth"],
        "description": "Get Details Of Current User using JWT token sent as cookie",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}
  